+++ CHANGE_LOG.md
@@ -6,6 +6,17 @@ All notable changes to this project will be documented in this file.
 
 ## [Unreleased]
 
+### Added
+
+- Padding property to control the padding size choice input adaptive card form field
+
+### Fixed
+
+- Audio button visibility state is tied to audio mute state
+- Restriction of the elements allowed to be render in the UI, to avoid security vulnerabilities.
+- New Error Event to notify CX regarding Error in widget
+- Padding property to control the padding size choice input adaptive card form field
+
 ## [1.7.3] 2024-11-20
 
 ### Added